requirements.txt
requests==2.28.2
websocket_client==1.6.1
